requirements.txt
asgiref==3.2.7
Django==2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pytz==2019.3
sqlparse==0.3.1
